:bug: get the right talks when multiple month are selected
