--- ozkour-back/domain/talks.js
@@ -1,35 +1,0 @@
-const dayjs = require('dayjs')
-const { getTalkFromDate } = require('../infrastructure/googlesheets/sheets.js')
-const customParseFormat = require('dayjs/plugin/customParseFormat')
-dayjs.extend(customParseFormat)
-
-const getTalk = async (start, end = dayjs()) => {
-  const formatedDateStart = dayjs(start)
-  const formatedDateEnd = dayjs(end)
-  console.log(start)
-  if (formatedDateStart.format('MM/YYYY') === formatedDateEnd.format('MM/YYYY')) {
-    console.log('same')
-    const param = { start: formatedDateStart.format('DD/MM/YYYY'), end: formatedDateEnd.format('DD/MM/YYYY') }
-    const res = await getTalkFromDate(param)
-    console.log(res)
-    return res
-  } else {
-    let res = []
-    let tempDateStart = formatedDateStart
-    let tempDateEnd = formatedDateStart.add(1, 'month')
-    while (tempDateStart.format('MM/YYYY') !== formatedDateEnd.format('MM/YYYY')) {
-      const param = { start: tempDateStart.format('DD/MM/YYYY'), end: tempDateEnd.format('DD/MM/YYYY') }
-      res = res.concat(await getTalkFromDate(param))
-      tempDateStart = tempDateStart.add(1, 'month')
-      tempDateEnd = tempDateStart.add(1, 'month')
-    }
-    const param = { start: tempDateStart.format('DD/MM/YYYY'), end: formatedDateEnd.format('DD/MM/YYYY') }
-    res = res.concat(await getTalkFromDate(param))
-    console.log(res)
-    return res
-  }
-}
-
-module.exports = {
-  getTalk
-}
