+++ ozkour-back/infrastructure/googlesheets/sheets.js
@@ -2,7 +2,11 @@ const dayjs = require('dayjs')
 const wrapper = require('./sheetWrapper')
 const utils = require('../../Utils/dateUtils')
 const customParseFormat = require('dayjs/plugin/customParseFormat')
+const isSameOrAfter = require('dayjs/plugin/isSameOrAfter')
+const isSameOrBefore = require('dayjs/plugin/isSameOrBefore')
 dayjs.extend(customParseFormat)
+dayjs.extend(isSameOrAfter)
+dayjs.extend(isSameOrBefore)
 
 /**
  * Get all the talks between 2 dates
@@ -11,48 +15,55 @@ dayjs.extend(customParseFormat)
  */
 async function getTalkFromDate (params) {
   // const res = await connect.authMethode(getData, param)
-  const month = utils.convDateToMonthInLetter(params.start)
-  const year = dayjs(params.start, 'DD/MM/YYYY').format('YYYY')
-  const res = await wrapper.getTalks(month, year)
-  return convertArrayToObject(dateFilter(res, params.start, params.end))
+  let tempDateStart = params.start
+  const tempDateEnd = params.end
+  const nbMonths = tempDateEnd.diff(tempDateStart, 'month')
+  let res = []
+  for (let i = 0; i <= nbMonths; i++) {
+    const month = utils.convDateToMonthInLetter(tempDateStart)
+    const year = dayjs(tempDateStart, 'DD/MM/YYYY').format('YYYY')
+    const test = await wrapper.getTalks(month, year)
+    res = res.concat(test)
+    tempDateStart = tempDateStart.add(1, 'month')
+  }
+  return res
 }
 
-/**
- * Filter the talks between 2 dates
- * @param {Array} talks the talks that need to be filtered
- * @param {String} start the start of the date range (format='DD-MM-YYYY')
- * @param {String} end the end of the date range (format='DD-MM-YYYY')
- */
-function dateFilter (talks, start, end) {
-  const formatedDateStart = dayjs(start, 'DD-MM-YYYY')
-  const formatedDateEnd = dayjs(end, 'DD-MM-YYYY')
-  talks = talks.filter(function (talk) {
-    const dateTalk = dayjs(talk[4], 'DD-MM-YYYY')
-    return formatedDateStart.isBefore(dateTalk.add(1, 'day'), 'day') && formatedDateEnd.isAfter(dateTalk.add(-1, 'day'), 'day')
-  })
-  return talks
-}
+// /**
+//  * Filter the talks between 2 dates
+//  * @param {Array} talks the talks that need to be filtered
+//  * @param {String} start the start of the date range (format='DD-MM-YYYY')
+//  * @param {String} end the end of the date range (format='DD-MM-YYYY')
+//  */
+// function dateFilter (talks, start, end) {
+//   const formatedDateStart = dayjs(start, 'DD-MM-YYYY')
+//   const formatedDateEnd = dayjs(end, 'DD-MM-YYYY')
+//   talks = talks.filter(function (talk) {
+//     const dateTalk = dayjs(talk[4], 'DD-MM-YYYY')
+//     return dateTalk.isSameOrAfter(formatedDateStart, 'day') && dateTalk.isSameOrBefore(formatedDateEnd, 'day')
+//     // return formatedDateStart.isBefore(dateTalk.add(1, 'day'), 'day') && formatedDateEnd.isAfter(dateTalk.add(-1, 'day'), 'day')
+//   })
+//   return talks
+// }
 
-function convertArrayToObject (arrayOfTalksArray) {
-  const arrayOfTalksObject = []
-  arrayOfTalksArray.forEach(([, eventType, eventName, universe, date, speakers, talkTitle]) => {
-    const value = {
-      date,
-      universe,
-      eventType,
-      eventName,
-      talkTitle,
-      speakers,
-      checked: true
-    }
+// function convertArrayToObject (arrayOfTalksArray) {
+//   const arrayOfTalksObject = []
+//   arrayOfTalksArray.forEach(([, eventType, eventName, universe, date, speakers, talkTitle]) => {
+//     const value = {
+//       date,
+//       universe,
+//       eventType,
+//       eventName,
+//       talkTitle,
+//       speakers,
+//       checked: true
+//     }
 
-    arrayOfTalksObject.push(value)
-  })
-  return arrayOfTalksObject
-}
+//     arrayOfTalksObject.push(value)
+//   })
+//   return arrayOfTalksObject
+// }
 
 module.exports = {
-  getTalkFromDate,
-  dateFilter,
-  convertArrayToObject
+  getTalkFromDate
 }
