+++ ozkour-back/tests/dateUtils.test.js
@@ -16,7 +16,7 @@ describe('dateUtils', () => {
   })
   describe('Display date as words', () => {
     it('should return a month as words, given a month in number format', () => {
-      expect(dateUtils.convDateToMonthInLetter('11/03/2021')).toBe('mars')
+      expect(dateUtils.convDateToMonthInLetter('11/03/2021')).toBe('Mars')
     })
     it('should return a date as words, given a month in number format', () => {
       expect(dateUtils.displayFullDateWithWords('11/03/2021')).toBe('11 mars 2021')
